requirements.txt
absl-py==2.1.0
accelerate==1.2.0
accelerated-scan==0.2.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alabaster==0.7.16
alembic @ file:///croot/alembic_1729532425530/work
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
apex @ file:///home/tabane/Documents/nemo/apex
appdirs==1.4.4
arpa==0.1.0b4
asciitree==0.3.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==5.0.1
attrdict==2.0.1
attrs==24.2.0
audioread==3.0.1
av==11.0.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
boto3==1.34.140
botocore==1.34.140
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
causal-conv1d==1.2.0.post2
cdifflib==1.2.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clip==0.2.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==5.0.1
comm==0.2.2
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
Cython @ file:///croot/cython_1727769790504/work
cytoolz==0.12.3
datasets==2.20.0
decorator==5.1.1
decord==0.6.0
diffusers==0.29.2
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0.post1
fasteners==0.19
faster-whisper==1.1.0
fasttext==0.9.3
fiddle==0.3.0
filelock==3.16.1
Flask==2.2.5
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.26.5
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
hypothesis==6.105.1
idna==3.10
ijson==3.3.0
imageio==2.34.2
imagesize==1.4.1
importlib_metadata==8.0.0
inflect==7.3.1
iniconfig==2.0.0
intervaltree==3.1.0
ipython==8.26.0
ipywidgets==8.1.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiwer==2.5.2
jmespath==1.0.1
joblib==1.4.2
julius==0.2.7
jupyterlab_widgets==3.0.11
kaldi-python-io==1.2.2
kaldiio==2.18.0
kenlm==0.2.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.4
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.22.0
lhotse==1.28.0
libcst==1.4.0
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities @ file:///home/conda/feedstock_root/build_artifacts/lightning-utilities_1732065998937/work
lilcom==1.8.0
llvmlite==0.43.0
loguru==0.7.2
lxml==5.2.2
Mako @ file:///croot/mako_1665472421453/work
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.21.3
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
megatron_core==0.5.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@0500d6b0f6e049a3ceb6bd2813de95d9be8fb4d1
nemo_text_processing==1.0.2
nerfacc==0.5.3
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numba==0.60.0
numcodecs==0.12.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.20.1
open_clip_torch==2.26.1
OpenCC==1.1.6
optuna==4.1.0
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pesq==0.0.4
pexpect==4.9.0
pillow==11.0.0
plac==1.4.3
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.0
primePy==1.3
progress==1.6
prompt_toolkit==3.0.47
propcache==0.2.1
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pyctcdecode==0.5.0
pydub==0.25.1
Pygments==2.18.0
pygtrie==2.5.0
pyloudnorm==0.1.1
PyMCubes==0.1.6
pynini==2.1.5
pyparsing==3.2.0
pypinyin==0.51.0
pypinyin-dict==0.8.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pystoi==0.4.1
pytest==8.2.2
pytest-mock==3.14.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorch-lightning @ file:///home/conda/feedstock_root/build_artifacts/pytorch-lightning_1723045326720/work
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==2.13.7
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.2
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==2.8.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.5.0
soxr==0.5.0.post1
speechbrain==1.0.2
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy @ file:///croot/sqlalchemy_1725885050993/work
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
textdistance==4.6.2
texterrors==0.5.1
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchdiffeq==0.2.4
torchmetrics @ file:///croot/torchmetrics_1719910137958/work
torchsde==0.2.6
torchvision==0.20.1
tqdm @ file:///croot/tqdm_1724853939799/work
traitlets==5.14.3
trampoline==0.1.2
transformers==4.47.0
trimesh==4.4.1
triton==3.1.0
typed-ast==1.5.5
typeguard==4.3.0
typer==0.15.1
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3==2.2.3
wandb==0.17.4
wcwidth==0.2.13
webdataset==0.2.86
Werkzeug==3.0.3
wget==3.2
whisperx @ git+https://github.com/federicotorrielli/BetterWhisperX.git@f7aa29c563e26d3b9072251910791f272ea67e67
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
yarl==1.18.3
youtokentome==1.0.6
yt-dlp==2024.12.6
zarr==2.18.2
zipp==3.19.2
